refactor: Exception 발생 시 예외 로그 찍는 부분을 Aop 로 분리 - week2
